pd-diplom/requirements.txt
Django==3.1
pytz==2019.1
django-rest-passwordreset==1.1.0
djangorestframework==3.11.1

celery==4.4.7
redis==3.5.3
requests==2.24.0
PyYAML==5.3.1
uritemplate==3.0.1
coreapi==2.3.3